requirements.txt
Pillow
pandas
python-docx
Flask
requests